Rewriting bot to use cogs
